Refactor: Rename the UnifiedMySiteMenu classes to drop the Unified and MySite words and move to a separate package under MySite
