+++ WordPress/src/main/AndroidManifest.xml
@@ -103,7 +103,7 @@
             android:exported="false" />
 
         <activity
-            android:name=".ui.mysite.UnifiedMySiteMenuActivity"
+            android:name=".ui.mysite.menu.MenuActivity"
             android:theme="@style/WordPress.NoActionBar"
             android:label="@string/my_site_section_screen_title"
             android:exported="false" />
