+++ WordPress/src/main/java/org/wordpress/android/ui/ActivityNavigator.kt
@@ -11,7 +11,7 @@ import org.wordpress.android.ui.blaze.blazecampaigns.campaignlisting.CampaignLis
 import org.wordpress.android.ui.blaze.blazepromote.ARG_BLAZE_FLOW_SOURCE
 import org.wordpress.android.ui.blaze.blazepromote.ARG_BLAZE_SHOULD_SHOW_OVERLAY
 import org.wordpress.android.ui.blaze.blazepromote.BlazePromoteParentActivity
-import org.wordpress.android.ui.mysite.UnifiedMySiteMenuActivity
+import org.wordpress.android.ui.mysite.menu.MenuActivity
 import org.wordpress.android.ui.mysite.personalization.PersonalizationActivity
 import javax.inject.Inject
 import javax.inject.Singleton
@@ -67,6 +67,6 @@ class ActivityNavigator @Inject constructor() {
     }
 
     fun openUnifiedMySiteMenu(context: Context) {
-        context.startActivity(Intent(context, UnifiedMySiteMenuActivity::class.java))
+        context.startActivity(Intent(context, MenuActivity::class.java))
     }
 }
