+++ WordPress/src/main/java/org/wordpress/android/ui/mysite/menu/MenuViewModel.kt
@@ -1,4 +1,4 @@
-package org.wordpress.android.ui.mysite
+package org.wordpress.android.ui.mysite.menu
 
 import androidx.lifecycle.MutableLiveData
 import dagger.hilt.android.lifecycle.HiltViewModel
@@ -8,6 +8,9 @@ import kotlinx.coroutines.flow.StateFlow
 import org.wordpress.android.fluxc.model.SiteModel
 import org.wordpress.android.modules.BG_THREAD
 import org.wordpress.android.ui.jetpack.JetpackCapabilitiesUseCase
+import org.wordpress.android.ui.mysite.MySiteCardAndItem
+import org.wordpress.android.ui.mysite.SelectedSiteRepository
+import org.wordpress.android.ui.mysite.SiteNavigationAction
 import org.wordpress.android.ui.mysite.items.listitem.SiteItemsBuilder
 import org.wordpress.android.ui.mysite.items.listitem.SiteItemsViewModelSlice
 import org.wordpress.android.ui.mysite.tabs.MySiteTabType
@@ -17,12 +20,12 @@ import org.wordpress.android.viewmodel.Event
 import org.wordpress.android.viewmodel.ScopedViewModel
 import javax.inject.Named
 
-data class UnifiedMenuViewState(
+data class MenuViewState(
     val items: List<MySiteCardAndItem> // cards and or items
 )
 
 @HiltViewModel
-class UnifiedMySiteMenuViewModel @Inject constructor(
+class MenuViewModel @Inject constructor(
     private val selectedSiteRepository: SelectedSiteRepository,
     @param:Named(BG_THREAD) private val bgDispatcher: CoroutineDispatcher,
     private val siteItemsBuilder: SiteItemsBuilder,
@@ -32,9 +35,9 @@ class UnifiedMySiteMenuViewModel @Inject constructor(
     private val _onNavigation = MutableLiveData<Event<SiteNavigationAction>>()
     val navigation = merge(_onNavigation, siteItemsViewModelSlice.onNavigation)
 
-    private val _uiState = MutableStateFlow(UnifiedMenuViewState(items = emptyList()))
+    private val _uiState = MutableStateFlow(MenuViewState(items = emptyList()))
 
-    val uiState: StateFlow<UnifiedMenuViewState> = _uiState
+    val uiState: StateFlow<MenuViewState> = _uiState
 
     fun start() {
         val site = selectedSiteRepository.getSelectedSite()!!
@@ -42,7 +45,7 @@ class UnifiedMySiteMenuViewModel @Inject constructor(
     }
 
     private fun buildSiteMenu(site: SiteModel) {
-        _uiState.value = UnifiedMenuViewState(
+        _uiState.value = MenuViewState(
             items = siteItemsBuilder.build(
                 siteItemsViewModelSlice.buildItems(
                     defaultTab = MySiteTabType.SITE_MENU,
@@ -59,7 +62,7 @@ class UnifiedMySiteMenuViewModel @Inject constructor(
     private fun updateSiteItemsForJetpackCapabilities(site: SiteModel) {
         launch(bgDispatcher) {
             jetpackCapabilitiesUseCase.getJetpackPurchasedProducts(site.siteId).collect {
-                _uiState.value = UnifiedMenuViewState(
+                _uiState.value = MenuViewState(
                     items = siteItemsBuilder.build(
                         siteItemsViewModelSlice.buildItems(
                             defaultTab = MySiteTabType.SITE_MENU,
